Splits server and client packages
- introduces libsheriff
- adds tests
