+++ libsheriff/.gitignore
@@ -1,0 +1,8 @@
+### Rust template
+# Generated by Cargo
+# will have compiled files and executables
+/target/
+
+# These are backup files generated by rustfmt
+**/*.rs.bk
+
