+++ server/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "move-the-sheriff-server"
+version = "1.0.0"
+authors = ["Austin Pray <austin@austinpray.com>"]
+edition = "2018"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
